requirements.txt
# Core AI and LLM frameworks
openai==1.54.3
langchain==0.3.7
langchain-openai==0.2.5
langchain-community==0.3.5
langchain-core==0.3.15

# Vector Store (starting with ChromaDB for ease of use)
chromadb==0.5.15

# Data validation and models
pydantic==2.9.2
pydantic-settings==2.6.1

# Image processing
Pillow==10.4.0

# Utilities
python-dotenv==1.0.1

# Optional: Observability and tracing
langfuse==2.52.1

# Testing and development
pytest==8.3.3
pytest-asyncio==0.24.0

# Type checking (optional)
mypy==1.13.0
